pyproject.toml
[project]
name = "langsearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=26.3.0",
    "isort>=8.0.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.22",
]
